move to _decorator package
