+++ dtoolkit/_decorator/__init__.py
@@ -1,0 +1,1 @@
+from .warning import warning  # noqa
